+++ src/ais/mod.rs
@@ -34,6 +34,7 @@ pub(crate) mod vdm_t19;
 pub(crate) mod vdm_t20;
 pub(crate) mod vdm_t21;
 pub(crate) mod vdm_t22;
+pub(crate) mod vdm_t23;
 pub(crate) mod vdm_t24;
 pub(crate) mod vdm_t27;
 
@@ -51,6 +52,7 @@ pub use vdm_t17::DgnssBroadcastBinaryMessage;
 pub use vdm_t20::{DataLinkManagementMessage};
 pub use vdm_t21::{AidToNavigationReport, NavAidType};
 pub use vdm_t22::{ChannelManagement};
+pub use vdm_t23::{GroupAssignmentCommand};
 
 // -------------------------------------------------------------------------------------------------